Fixed file locations for recent parameter updates
